requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Graph DB
neo4j==5.16.0

# Vector DB
chromadb==0.4.22

# AI
anthropic==0.40.0

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
pytesseract==0.3.10

# Data
pandas==2.1.4
numpy==1.26.3

# Security / Backup
cryptography==42.0.0

# CLI
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
